CVSL-4096 consume prisoner updated event for restricted patients
